Tests: introduce an `SslTestCase` class
... containing various methods which are used by tests for multiple different methods.
